net: hardware capabilities discovery - matches and actions
This patch adds infrastructure for network hardware devices to
expose the headers and fields they can parse along with the
actions they can perform on packets.

This work started initially as an attempt to extend the flow
classifier interface which is part of ethtool. the ethtool
interface currently suffers from a few problems we wish to
correct. First it does not have any mechanism to expose supported
matches/actions. This limitation makes it difficult to build
intelligent controllers. A typical problem we encounter is placing
workloads on systems that can meet a SLA (service level agreement)
which may include hardware acceleration/resources. This is common
in SR-IOV use cases where we may only want a direct assigned VF
if it supports some set of actions and parsing on specific fields.
This patch is the first step in resolving this by exposing a set
of standard actions and matches along with a new netlink family
to export them.

Another issue we discovered while trying to extend ethtool is
because it is ioctl based it does not support multicast and
broadcast features similar to netlink. By moving the parser and
action capabilities over to netlink we get this for free. This
is very useful when multiple agents are trying to manage and/or
discover hardware devices.

The last issue we try to address here is flexibility. By exposing
the parser using a generic bit:offset structure we can easily add
new fields and protocols. This facilitates hardware that can be
used to match on application specific protocols.

Finally although this patch series only exposes read only support
there are devices that support programming the parser. The header
graph exposed here along with the bit:offset format would allow
programming a hardware device with a new parser. At the moment we
do not expect programming new actions would be feasible due to
difficulty creating well-defined actions when they are generated
by the driver.

This work is based on an initial submission and presentation
at netdev0.1 back in Feb. called the Flow API. We are working
to address a series of critique which we roughly believe to
be based on (1) actions being exposed by hardware and (2) programming
hardware directly through a new API instead of using existing
APIs. The reviewer of the previous series will notice this
series drops the add_flow, del_flow rules and also no longer
allows hardware to define action sets.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>