net: hardware capabilities discovery - report tables
Packet processing pipelines implemented in hardware can be modelled
as a series of tables. We see this today with the standard software
tables such as, fdb, fib, u32-tc, bpf-maps, etc.

This provides a mechanism for drivers to expose these tables and
the capabilities of each table. For table has the following properties

	- max number of elements
	- fields it can match on
	- actions it can perform
	- subsystem binding
	- subsystem arguments

The max number of elements gives the capacity of the table. This
allows controllers to calculate when fallback operations will be
performed. For example if the L2 table has n rules in it we can
learn that adding an n+1 rule will result in either the rule being
rejects or dropping back to software paths depending on error
handling code.

The match fields is used to determine what fields the hardware
can support matching on. Extending the L2 table example we may
support hardware that can match on both MAC and VLAN but also
hardware that can only match on MAC address. Both types of hardware
exists. A controller may query the hardware for minimal support
and avoid placing workloads on hardware that doesn't meet the more
stringent MAC/VLAN requirement if needed.

The actions it can perform identifies what types of actions are
supported by the device. For basic L2 tables we would minimally
expect a forward action. But when we try to offload 'tc' filters
and actions we will need to have a mechanism to check for specific
actions.

The subsystem binding tells us which part of the networking stack
is managing the device. This patch exposes three initial bindings
L2, L3, and bpf-map. When 'L2' is indicated we can expect the
table to be hooked into the 'ndo_fdb*' operations. Similarly, with
a L3 binding we expect the routing subsystem to be configuring the
table. Finally, the 'bpf-map' binding can be used to indicate the
table is being kept in sync with a bpf-map. This mode "reflects"
any entry in the map into the hardware and vice versa indicates
the hardware may update the map.

At the moment the arguments entry is only used with the bpf-map
to indicate the 'fd' of the map that the table is being "reflected"
from. See subsequent patches for the details.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>