net: hardware provisioning - create/destroy table
Now that hardware tables can be reported using get_tables command we
may want to customize the pipeline for a paticular use case. Many of
the hardware devices we are aware of are built on top of a set of
TCAMs, hash tables, etc data structures that can be allocated to
different functions at init time or even at run time.

One of the most basic examples we are aware of is the fm10k device
which has TCAM that can be used to do augment the L2 or L3 tables
(both of which currently support pushing rules into hardware) or
possibly mapped into other subsystems in the networking stack.
Further this hardware resource can be partitioned and divided amongst
different subsystems. For example the user may use some portion of
the resource to augment the L2 table, another piece to augment the
L3 stack, and finally use the rest to reflect a BPF map table into the
hardware.

Also, although we are working on the fm10k device in this series
primarily because the authors have access to this NIC we are aware
of other devices which have similar shared resources that are can
be allocated. In fact we think this is a fairly common feature of
modern NIC and/or switch devices.

This patch adds the create table and destroy table interface.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>