net: share bpf map with allocated hardware table
This adds the necessary hooks for a BPF map to be kept in sync with a
device driver. Here every time an entry is added to the BPF map the
entry is also added to the hardware map in this way the software and
hardware map are 1:1 in sync. Similarly when an entry is updated or
deleted the corresponding call is made on the device driver side to
sync the maps.

todo:
 - could put some build flags to only build when drivers with support
   are being built

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>