net: hardware provisioning - support bpf table maps
With the BPF hooks added to sync a hardware table with a software map
this patch enables the flow_table generic bits to do the 'binding' and
'unbind' of the table. Additionally at bind time the hardware is put
in a 1:1 sync with the map.

We have found that it is useful to have software tables operating
normally and then to bind them and unbind them with hardware
dynamically without disrupting the software maps. This is supported
by providing the map 'fd' with the create table operation.

todo:
	- we don't seem to have the map type included here which
	  is likely a problem at the moment I assume its a hash in
	  the driver.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>